+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
 
-## [1.3.19] - Unreleased
+## [1.3.19] - 2022-11-01
 - Add ColorChooser::set_rgb().
 - Add check in MenuExt::text() for empty menus by @purrie.
 - Add other similar checks for other MenuExt methods.
+- Impl PartialEq, Eq and Default for StyleTableEntry and StyleTableEntryExt.
 - Pull FLTK fixes.
 
 ## [1.3.18] - 2022-10-21